docugen-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
openai = "^1.99.7"
elevenlabs = "^2.9.2"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
requests = "^2.31.0"
google-api-python-client = "^2.108.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
pillow = ">=9.0.0,<10.0.0"
moviepy = ">=1.0.3,<2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
